+++ packages/integrations/image/components/index.js
@@ -1,1 +1,2 @@
 export { default as Image } from './Image.astro';
+export { default as Picture } from './Picture.astro';